requirements.txt
# requirements.txt
flask
python-docx
reportlab
pandas